Refs #9571: Restructured validation script into a class and shared files.
